ZEPPELIN-3974. Reorgnize zeppelin integration test
